refactor(infra): add migration logic
