+++ packages/infra/src/__tests__/workspace.ydoc
