+++ packages/infra/tsconfig.node.json
@@ -8,5 +8,13 @@
     "outDir": "lib",
     "noEmit": false
   },
-  "include": ["vite.config.ts"]
+  "include": ["vite.config.ts", "src/__tests__"],
+  "references": [
+    {
+      "path": "./tsconfig.json"
+    },
+    {
+      "path": "../../tests/fixtures"
+    }
+  ]
 }
